# application/cmd/server/docs/swagger.yaml
basePath: /api/v1
definitions:
  Kullanicilar.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  Kullanicilar.LoginRequest:
    properties:
      kullaniciAdi:
        type: string
      sifre:
        type: string
    required:
    - kullaniciAdi
    - sifre
    type: object
  Kullanicilar.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  entities.CreateUserRequest:
    properties:
      dogumTarihi:
        description: YYYY-MM-DD formatında
        type: string
      eposta:
        type: string
      isim:
        type: string
      kullaniciAdi:
        maxLength: 32
        minLength: 3
        type: string
      sifre:
        minLength: 6
        type: string
      soyisim:
        type: string
      tcKimlikNo:
        type: string
      telefon:
        type: string
    required:
    - dogumTarihi
    - eposta
    - isim
    - kullaniciAdi
    - sifre
    - soyisim
    - tcKimlikNo
    - telefon
    type: object
...
info:
  contact: {}
  description: This is the Ihale API server.
  title: Ihale API
  version: "1"
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: Logs in a user with a username and password.
      parameters:
      - description: Login Request
        in: body
        name: loginRequest
        required: true
        schema:
          $ref: '#/definitions/Kullanicilar.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: login successful
          schema:
            type: string
        "400":
          description: Request parsing failed
          schema:
            $ref: '#/definitions/Kullanicilar.ErrorResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/Kullanicilar.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/Kullanicilar.ErrorResponse'
      summary: User login
      tags:
      - kullanicilar
  /signup:
    post:
      consumes:
      - application/json
      description: Yeni bir kullanıcı kaydeder.
      parameters:
      - description: Kayıt Bilgileri
        in: body
        name: signUpRequest
        required: true
        schema:
          $ref: '#/definitions/entities.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Kullanıcı başarıyla kaydedildi
          schema:
            $ref: '#/definitions/Kullanicilar.MessageResponse'
        "400":
          description: İstek ayrıştırması başarısız
          schema:
            $ref: '#/definitions/Kullanicilar.ErrorResponse'
        "500":
          description: Ledger üzerinde kullanıcı kaydedilemedi
          schema:
            $ref: '#/definitions/Kullanicilar.ErrorResponse'
      summary: Kullanıcı Kaydı
      tags:
      - kullanicilar
...
swagger: "2.0"